backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.0
python-dotenv==1.0.0

# Google Gemini AI and LangGraph
google-generativeai>=0.7.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# Yahoo Finance and Stock Data
yfinance==0.2.32
yahooquery==2.3.7

# Vector database
chromadb==0.4.18

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3

pydantic-settings==2.11.0

